+++ smartcosmos-extension-api/src/main/java/net/smartcosmos/platform/constraint/validator/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Collection of validators for custom validation annotations.
+ */
+package net.smartcosmos.platform.constraint.validator;